requirements.txt
intersight
python-dotenv
requests
urllib3
intersight-auth